move filtered informer and namespace filter to pkg.kube
