+++ pilot/pkg/serviceregistry/kube/controller/fake.go
@@ -20,10 +20,10 @@ import (
 	meshconfig "istio.io/api/mesh/v1alpha1"
 	"istio.io/istio/pilot/pkg/model"
 	"istio.io/istio/pilot/pkg/serviceregistry/aggregate"
-	"istio.io/istio/pilot/pkg/serviceregistry/kube/controller/filter"
 	"istio.io/istio/pkg/cluster"
 	"istio.io/istio/pkg/config/mesh"
 	kubelib "istio.io/istio/pkg/kube"
+	filter "istio.io/istio/pkg/kube/namespace"
 	"istio.io/istio/pkg/test"
 )
