+++ pkg/kube/informer/filter.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package filter
+package informer
 
 import (
 	"k8s.io/client-go/tools/cache"
