+++ pkg/kube/namespace/filter.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package filter
+package namespace
 
 import (
 	"sync"
